requirements.txt
asgiref==3.6.0
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
secret-key-generator==0.0.8
slippers==0.5.0
sqlparse==0.4.3
